Reset DA phase recovery state by default
Change the parameter that defines whether or not to reset the internal state of
the data-aided phase recovery loop after every frame. Default to reset the state
on every frame.

The motivation is that this effectively makes use of the data-aided phase error
and fine frequency offset estimations that are computed by the new frame
synchronizer. The frame sync block computes these metrics on every frame and
places this information within stream tags that are read by the phase recovery
block. The latter, in turn, only uses this information when configured to reset
state on every frame. When this is the case, the phase recovery block
re-initializes its control loop with both phase and fine frequency offset
estimation values. The phase becomes the initial phase of its NCO and the
frequency offset becomes the initial value of the integrator. This speeds up
convergence of the phase recovery loop during the preamble.